requirements.txt
alembic==1.4.1
BareNecessities==0.2.8
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
click==7.1.1
cryptography==2.8
dominate==2.5.2
Flask==1.1.1
Flask-Admin==1.5.6
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
jwt==0.6.1
Mail==2.1.0
Mako==1.1.2
MarkupSafe==1.1.1
migrate==0.3.8
olefile==0.46
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
six==1.14.0
SQLAlchemy==1.3.15
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
